requirements.txt
# all vendored requirements
-r vendor/requirements.txt

# all extension requirements


# all AlbertoX3-core requirements
aenum~=3.1.11
aiomysql~=0.1.1
interactions[speedup,sentry] @ git+https://github.com/interactions-py/interactions.py.git@5.x
python-datetimeparser @ git+https://github.com/aridevelopment-de/datetimeparser.git@dev
python-dotenv~=1.0.0
PyYAML~=6.0
redis~=4.5.3
sqlalchemy~=2.0.7
